Minor test cleanup
